Fix using datetime in custom template (#83)
Revert f52916076089568abe5294502042df001e43ce5d

+++ main.go
@@ -190,12 +190,12 @@ func main() {
 			Usage:  "pull request",
 			EnvVar: "DRONE_PULL_REQUEST",
 		},
-		cli.Int64Flag{
+		cli.Float64Flag{
 			Name:   "job.started",
 			Usage:  "job started",
 			EnvVar: "DRONE_BUILD_STARTED",
 		},
-		cli.Int64Flag{
+		cli.Float64Flag{
 			Name:   "job.finished",
 			Usage:  "job finished",
 			EnvVar: "DRONE_BUILD_FINISHED",
@@ -281,8 +281,8 @@ func run(c *cli.Context) error {
 			Event:    c.String("build.event"),
 			Status:   c.String("build.status"),
 			Link:     c.String("build.link"),
-			Started:  c.Int64("job.started"),
-			Finished: c.Int64("job.finished"),
+			Started:  c.Float64("job.started"),
+			Finished: c.Float64("job.finished"),
 			PR:       c.String("pull.request"),
 			DeployTo: c.String("deploy.to"),
 		},
